+++ tests/interface_dag_elements/test_warnings.py
@@ -151,6 +151,7 @@ def test_warn_if_evaluation_date_set_in_multiple_places_implicitly_added(backend
             policy_environment=policy_environment,
             evaluation_date=datetime.date(2025, 1, 1),
             processed_data={"p_id": xnp.array([0])},
+            input_data=InputData.tree(tree={"p_id": xnp.array([0])}),
             tt_targets=TTTargets.tree({"p_id": None}),
             backend=backend,
         )
@@ -171,6 +172,7 @@ def test_do_not_need_to_warn_if_evaluation_date_is_set_only_once(backend, xnp):
             policy_environment=policy_environment,
             evaluation_date=datetime.date(2025, 1, 1),
             processed_data={"p_id": xnp.array([0])},
+            input_data=InputData.tree(tree={"p_id": xnp.array([0])}),
             tt_targets=TTTargets.tree({"p_id": None}),
             backend=backend,
         )